init/nitro: Add shutdown FD for net proxy dealloc
Originally; when the enclave application ended, there would be no
explicit closing of the vsock file descriptor within the enclave network
proxy. With this, the host network proxy process never ended and was
essentially a zombie.

When the application ends, send a signal (via an eventfd) to the guest
network proxy, which will read this signal and explicitly close the
proxy file descriptors. This will signal to the krun-nitro proxy that
the guest proxy has ended and it itself should close the file
descriptors and exit as well.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>