nitro/net: Signal passt free on guest shutdown
When the host network proxy receives the "shutdown signal" from the
guest's vsock, it must signal to the passt listener thread that it must
also end the stream and exit.

Add a sender/receiver channel in which the host network vsock listener
notifies the passt listener to exit before it itself exits. Upon
receiving this message, the passt proxy can then exit.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>